Move BOOKS const from its own file to src/index.js
