+++ src/index.js
@@ -6,7 +6,25 @@ import './index.css';
 import App from './components/App';
 import rootReducer from './reducers/index';
 
-const store = createStore(rootReducer);
+const BOOKS = [
+  {
+    id: Math.floor(Math.random() * 10000),
+    title: 'Book Test 1',
+    category: 'Action',
+  },
+  {
+    id: Math.floor(Math.random() * 10000),
+    title: 'Book Test 2',
+    category: 'Horror',
+  },
+  {
+    id: Math.floor(Math.random() * 10000),
+    title: 'Book Test 3',
+    category: 'Sci-Fi',
+  },
+];
+
+const store = createStore(rootReducer, { books: BOOKS });
 
 ReactDOM.render(
   <React.StrictMode>
