+++ src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import bookReducer from './books';
 
-const rootReducer = combineReducers({ bookReducer });
+const rootReducer = combineReducers({ books: bookReducer });
 
 export default rootReducer;
