--- src/states/index.js
@@ -1,19 +1,0 @@
-const BOOKS = [
-  {
-    id: Math.floor(Math.random() * 10000),
-    title: 'Book Test 1',
-    category: 'Action',
-  },
-  {
-    id: Math.floor(Math.random() * 10000),
-    title: 'Book Test 2',
-    category: 'Horror',
-  },
-  {
-    id: Math.floor(Math.random() * 10000),
-    title: 'Book Test 3',
-    category: 'Sci-Fi',
-  },
-];
-
-export default BOOKS;
